+++ probe-rs/src/architecture/arm/core/armv7m.rs
@@ -637,6 +637,10 @@ impl<'probe> Armv7m<'probe> {
             sequence,
         })
     }
+
+    fn set_core_status(&mut self, new_status: CoreStatus) {
+        super::update_core_status(&mut self.memory, &mut self.state.current_state, new_status);
+    }
 }
 
 impl<'probe> CoreInterface for Armv7m<'probe> {
@@ -667,7 +671,7 @@ impl<'probe> CoreInterface for Armv7m<'probe> {
                 "The core is in locked up status as a result of an unrecoverable exception"
             );
 
-            self.state.current_state = CoreStatus::LockedUp;
+            self.set_core_status(CoreStatus::LockedUp);
 
             return Ok(CoreStatus::LockedUp);
         }
@@ -678,7 +682,7 @@ impl<'probe> CoreInterface for Armv7m<'probe> {
                 tracing::warn!("Expected core to be halted, but core is running");
             }
 
-            self.state.current_state = CoreStatus::Sleeping;
+            self.set_core_status(CoreStatus::Sleeping);
 
             return Ok(CoreStatus::Sleeping);
         }
@@ -710,7 +714,7 @@ impl<'probe> CoreInterface for Armv7m<'probe> {
                 );
             }
 
-            self.state.current_state = CoreStatus::Halted(reason);
+            self.set_core_status(CoreStatus::Halted(reason));
 
             return Ok(CoreStatus::Halted(reason));
         }
@@ -720,7 +724,7 @@ impl<'probe> CoreInterface for Armv7m<'probe> {
             tracing::warn!("Core is running, but we expected it to be halted");
         }
 
-        self.state.current_state = CoreStatus::Running;
+        self.set_core_status(CoreStatus::Running);
 
         Ok(CoreStatus::Running)
     }
@@ -786,7 +790,7 @@ impl<'probe> CoreInterface for Armv7m<'probe> {
         self.memory.flush()?;
 
         // We assume that the core is running now
-        self.state.current_state = CoreStatus::Running;
+        self.set_core_status(CoreStatus::Running);
 
         Ok(())
     }