+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added support for the Olimex ARM-USB-TINY-H JTAG device (#1586).
+- Added support for propagating `CoreStatus` to the probe in use (#1588).
 
 ## [0.18.0]